fix: 발표방 생성 수정

+++ src/main/java/com/oronaminc/join/room/dto/RoomDetailResponse.java
@@ -17,7 +17,7 @@ public record RoomDetailResponse(
         String documentUrl,
         Integer participantCount,
         Integer participantLimit,
-        Integer emojiCount,
+        Long emojiCount,
         boolean isHost,
         boolean isTeamMember,
         RoomStatus roomStatus,
